Split composer pipeline by responsibility
